Implement tweeting functionality
This is split in 2 parts: Tweeting and universalization

1. When a user is live, the StreamEventHandler calls the TweetHelper
to publish a pre-made tweet that the user is live with the title
of his or her stream and the twitch URL, just like it would appear
on a twitch going live tweet!

2. Add the ability to add new creators via a json file. The usernames
file is read into a ConcurrentDictionary since this is a thread safe
collection. When the stream goes live, the TweetHelper looks for the
provided twitch username in the dictionary and get's the user's twitter
username from it. Now, we can mention the twitter user and provide
the twitch link at the same time with this simple logic.

Signed-off-by: kingbri <bdashore3@gmail.com>