Add a system for cooldowns
A streamer may accidentally go offline due to connection issues
or attempt to reconnect to twitch servers by restarting the stream.
If this remained unchecked, tweets would blow up on the account every
single time the twitch API picks up that the user has streamed.

Provide a simple and elegant solution to this problem by introducing
a local cooldown timer. Once Twitch sends the signal that the stream
is live, store the current UNIX time in a concurrent dictionary along
with the Twitch username of the streamer.

When the streamer goes offline, the bot will remove the stored
time to free up memory used by the dictionary key/pair value.

If the user accidentally goes offline within one and a half hours,
the cooldown will prevent the removal of the key/pair value and break
out of the function.

If the user ends the stream after one and a half hours, the key/pair
value will stay until the user goes live again or the bot is manually
restarted. (I need to implement commands in the bot for manual use).

In the end, this is a simple solution that covers almost all the bases
of this problem.

Signed-off-by: kingbri <bdashore3@gmail.com>