Update with async methods and add a test
Use Tweetinvi's asynchronous methods for posting tweets when
a user goes live.

Also, there was a fatal error when running the last iteration of
the bot in my server. It kept crashing when checking if the key
exists in the dictionary when checking for cooldown. Therefore, I
need to add another check to see if the key exists. Keep this Console
statement in there for debugging over the night

This commit will be reformatted once I fix this issue

Signed-off-by: kingbri <bdashore3@gmail.com>